Update patch set 1
Patch Set 1: Code-Review-1

You appear to be trying to use the Builder Pattern, however the implementation does not conform to one, since the builder class does not have a build method that returns the object you are trying to build.

What you have actually implemented appears to be a POJO-like class with special setters that returns itself, and using that to pass multiple values into a method call.

I recommend that you either convert your new class into a regular POJO and use it as a simple data object which wraps the input parameters, or implement the builder pattern in earnest to create a builder that can be used to build a SpikeEventEntity object directly.

Patch-set: 1
Reviewer: Gerrit User 703 <703@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1