Update patch set 3
Patch Set 3: Code-Review-1

We currently do not have the original supporters of this microservice supporting it, with just the minimal set of test cases passing we are not comfortable with merging such large changes that could potentially cause issues. If you were able to build and run the docker image to prove the original functionality is intact then we will merge it. Currently if this was merged and caused issues in our release we do not have the personnel to address it.

Patch-set: 3
Reviewer: Gerrit User 2979 <2979@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1